+++ src/reexp.rs
@@ -1,0 +1,7 @@
+pub use cpal::{
+    BuildStreamError, DefaultStreamConfigError, DevicesError, FromSample,
+    PauseStreamError, PlayStreamError, SampleFormat, StreamError,
+    SupportedStreamConfigsError,
+};
+
+pub use anyhow::Result as AnyhowResult;