+++ src/shared_data.rs
@@ -1,13 +1,12 @@
 use std::{
     fmt::Debug,
     sync::{Mutex, MutexGuard},
-    time::{Duration, Instant},
 };
 
-use crate::{Error, callback::Callback, err::Result, source::Source};
+use crate::{Callback, CallbackInfo, Controls, Error, Result, Source};
 
 /// Data shared between sink and the playback loop
-pub(super) struct SharedData {
+pub struct SharedData {
     /// Used to control the playback loop from the [`Sink`]
     controls: Mutex<Controls>,
     /// The source for the audio
@@ -16,27 +15,6 @@ pub(super) struct SharedData {
     callback: Callback<CallbackInfo>,
     /// Function used as callback when errors occur on the playback loop
     err_callback: Callback<Error>,
-}
-
-/// Used to control the playback loop from the sink
-#[derive(Clone, Debug)]
-pub(super) struct Controls {
-    /// Fade duration when play/pause
-    pub(super) fade_duration: Duration,
-    /// When true, playback plays, when false playback is paused
-    pub(super) play: bool,
-    /// Sets the volume of the playback
-    pub(super) volume: f32,
-}
-
-/// Callback type and asociated information
-#[non_exhaustive]
-#[derive(Debug)]
-pub enum CallbackInfo {
-    /// Invoked when the current source has reached end
-    SourceEnded,
-    /// Invoked when no sound is playing and you can call hard_pause
-    PauseEnds(Instant),
 }
 
 impl SharedData {
@@ -96,22 +74,5 @@ impl Debug for SharedData {
             .field("callback", &self.callback)
             .field("err_callback", &self.err_callback)
             .finish()
-    }
-}
-
-impl Controls {
-    /// Creates new controls
-    pub(super) fn new() -> Self {
-        Self {
-            fade_duration: Duration::ZERO,
-            play: false,
-            volume: 1.,
-        }
-    }
-}
-
-impl Default for Controls {
-    fn default() -> Self {
-        Self::new()
     }
 }