+++ src/source/symph/options.rs
@@ -1,0 +1,6 @@
+pub use symphonia::core::formats::FormatOptions;
+
+#[derive(Debug, Default)]
+pub struct Options {
+    pub format: FormatOptions,
+}